+++ Components/Screens/DrawerNavigator.js
@@ -2,13 +2,13 @@ import { createDrawerNavigator } from 'react-navigation';
 
 import HomeScreen from './Home'
 import Profile from './Profile'
-import Flights from './Flights'
+import BookFlights from './BookFlights'
 import Test from './Test2'
 import LogOut  from './LogOut'
 
 export default createDrawerNavigator({
   Home: { screen: HomeScreen },
-  Flights: { screen: Flights },
+  Flights: { screen: BookFlights },
   Test: { screen: Test },
   Profile: { screen: Profile },
   LogOut: { screen: LogOut },
