+++ Components/Screens/Home.js
@@ -1,8 +1,25 @@
 import React, { Component }from 'react';
-
-import { Icon, Content, Container, Header, Button, Left, Right, Body, Text } from 'native-base';
+import { Image } from 'react-native';
+import { Icon, Container, Header, Button, Left, Right, Body, Text, DeckSwiper, Card, CardItem } from 'native-base';
 
 import GlobalStyles from '../../GlobalStyles';
+
+
+const cards = [
+  {
+    name: 'Oferta 1',
+    image: require('../Images/oferta_ficticia_1.png'),
+  },
+  {
+    name: 'Oferta 2',
+    image: require('../Images/oferta_ficticia_2.png'),
+  },
+  {
+    name: 'Oferta 3',
+    image: require('../Images/oferta_ficticia_3.png'),
+  }
+];
+
 
 export default class HomeScreen extends Component {
     static navigationOptions = {
@@ -29,11 +46,22 @@ export default class HomeScreen extends Component {
               </Button>
             </Right>
           </Header>
-          <Content contentContainerStyle={GlobalStyles.container}>
-            <Text>
-              Home Content
-            </Text>
-          </Content>
+
+          <DeckSwiper
+            dataSource={cards}
+            renderItem={item =>
+              <Card style={{ elevation: 3 }}>
+                <CardItem cardBody>
+                  <Image style={{ height: 300, flex: 1 }} source={item.image} />
+                </CardItem>
+                <CardItem>
+                  <Icon name="heart" style={{ color: '#ED4A6A' }} />
+                  <Text>{item.name}</Text>
+                </CardItem>
+              </Card>
+            }
+          />
+
         </Container>
       );
     }
